backend/pyproject.toml
[tool.poetry]
name = "ai-chat-backend"
version = "0.1.0"
description = "A simple chat API connecting to OpenAI"
authors = ["Your Name <your.email@example.com>"]
packages = [
    {include = "*.py"}
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
openai = "^1.7.0"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 